requirements.txt
django>=4.2.0
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv>=1.0.0